Ensure temporary results_dir is created
* The results_dir is created to save log files during tempest run
* Ensure it is created correctly

Closes-Bug: #1335122

+++ rally/benchmark/context/tempest.py
@@ -36,7 +36,12 @@ class Tempest(base.Context):
     @utils.log_task_wrapper(LOG.info, _("Enter context: `tempest`"))
     def setup(self):
         self.verifier = tempest.Tempest(self.task.task.deployment_uuid)
-        self.verifier.log_file = "/dev/null"
+        self.verifier.log_file_raw = "/dev/null"
+        # Create temporary directory for subunit-results.
+        self.results_dir = os.path.join(
+            tempfile.gettempdir(), "%s-results" % self.task.task.uuid)
+        os.mkdir(self.results_dir)
+        self.context["tmp_results_dir"] = self.results_dir
 
         try:
             if not self.verifier.is_installed():
@@ -54,12 +59,6 @@ class Tempest(base.Context):
 
         self.context["verifier"] = self.verifier
 
-        # Create temporary directory for xml-results.
-        self.results_dir = os.path.join(
-            tempfile.gettempdir(), "%s-results" % self.task.task.uuid)
-        os.mkdir(self.results_dir)
-        self.context["tmp_results_dir"] = self.results_dir
-
     @utils.log_task_wrapper(LOG.info, _("Exit context: `tempest`"))
     def cleanup(self):
         try:
@@ -74,4 +73,6 @@ class Tempest(base.Context):
                                   cwd=self.verifier.tempest_path)
         except subprocess.CalledProcessError:
             LOG.error("Tempest cleanup failed.")
-        shutil.rmtree(self.results_dir)
+
+        if os.path.exists(self.results_dir):
+            shutil.rmtree(self.results_dir)
